Initial Commit for NewJetControl
I've made the calculaltions a bit more logical.
Screenspace (x,y) selection is oriented to the viewport (z is forward).
Then the vector from character to target coord is projected onto the
viewport forward plane (this gives a direction from character to virtual
cursor in a plane parellel with the viewport passing through character).